refasctor: hooks folder added + minor refactoring | mz
