+++ src/hooks/useBinance.tsx
@@ -1,39 +1,18 @@
 import { useEffect, useState, useMemo } from 'react';
 import {
-	useLocalStorage,
 	BINANCE_EXCHANGE_INFO,
 	BINANCE_PRICE_TICKER,
 	isTokenSelected,
 	isNetworkSelected,
 	PROTOCOL_FEE_FACTOR
 } from '../helpers';
-import { DestinationNetworks } from '../styles';
+import { DestinationNetworks } from '../helpers';
 import destinationNetworks from '../data/destinationNetworks.json';
 import { useEtherBalance, useEthers } from '@usedapp/core';
 import { formatEther } from '@ethersproject/units';
-// eslint-disable-next-line camelcase
-import jwt_decode from 'jwt-decode';
-import dayjs from 'dayjs';
+import { useStore, START_TOKEN } from '../helpers';
+import type { Price } from '../helpers';
 import axios from 'axios';
-import {
-	INITIAL_STORAGE,
-	LOCAL_STORAGE_AUTH,
-	BASE_URL,
-	routes,
-	useStore,
-	VerificationEnum,
-	START_TOKEN
-} from '../helpers';
-
-type JwtType = {
-	iss: string;
-	exp: number;
-	type: string;
-};
-
-export const message = {
-	ignore: 'ignore'
-};
 
 type Ticker = {
 	baseAsset: string;
@@ -42,79 +21,7 @@ type Ticker = {
 	filters: any[];
 };
 
-export type Price = { symbol: string; price: string };
-
-export const useAxios = () => {
-	const [storage, setStorage] = useLocalStorage(LOCAL_STORAGE_AUTH, INITIAL_STORAGE);
-	const {
-		state: { accessToken, refreshToken },
-		dispatch
-	} = useStore();
-
-	const axiosInstance = axios.create({
-		baseURL: BASE_URL,
-		headers: { Authorization: `Bearer ${accessToken}` }
-	});
-
-	axiosInstance.interceptors.request.use(
-		async (req) => {
-			const access: JwtType = jwt_decode(accessToken);
-			const isAccessTokenExpired = dayjs.unix(access?.exp).diff(dayjs()) < 1;
-
-			if (!isAccessTokenExpired) return req;
-			// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-			req.headers!.Authorization = `Bearer ${refreshToken}`;
-
-			const refresh: JwtType = jwt_decode(refreshToken);
-			const isRefreshTokenExpired = dayjs.unix(refresh?.exp).diff(dayjs()) < 1;
-
-			if (!isRefreshTokenExpired) return req;
-
-			try {
-				const newTokens = await axios.post(
-					`${BASE_URL}${routes.refresh}`,
-					{},
-					{
-						headers: {
-							Authorization: `Bearer ${refreshToken}`,
-							'Content-Type': 'application/json',
-							'Access-Control-Allow-Origin': '*'
-						}
-					}
-				);
-
-				dispatch({ type: VerificationEnum.ACCESS, payload: newTokens.data.access });
-				dispatch({ type: VerificationEnum.REFRESH, payload: newTokens.data.refresh });
-				setStorage({ ...storage, access: newTokens.data.access, refresh: newTokens.data.refresh });
-				// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-				req.headers!.Authorization = `Bearer ${newTokens.data.access}`;
-
-				return req;
-			} catch (error: any) {
-				error.message = message.ignore;
-
-				return Promise.reject(error);
-			}
-		},
-		(error) => {
-			return Promise.reject(error);
-		}
-	);
-
-	axiosInstance.interceptors.response.use(
-		(res) => res,
-		(error) => {
-			// TODO: check for auth urls how to handle to not throw 401 fo refresh etc.
-			if (Object.keys(error).length === 1 && 'message' in error) error.message = message.ignore;
-
-			return Promise.reject(error);
-		}
-	);
-
-	return axiosInstance;
-};
-
-export const useBinanceApi = () => {
+export const useBinance = () => {
 	// TODO: error handling if API calls fail
 	const [allPrices, setAllPrices] = useState<Price[]>([]);
 	const [allFilteredPrices, setAllFilteredPrices] = useState<Price[]>([]);
