+++ src/hooks/useFees.tsx
@@ -1,0 +1,3 @@
+export const useFees = () => {
+	return 'true';
+};
